Move some classes into separate files (#897)
